[SG-70] modify: move signin vm in vm module
