+++ app/src/main/kotlin/com/captures2024/soongan/SoonGanActivity.kt
@@ -26,11 +26,11 @@ import com.captures2024.soongan.core.auth.kakao.KakaoAuthHelperImpl
 import com.captures2024.soongan.core.auth.kakao.KakaoLoginCallback
 import com.captures2024.soongan.core.designsystem.theme.SoonGanTheme
 import com.captures2024.soongan.core.viewmodel.AppRootViewModel
+import com.captures2024.soongan.core.viewmodel.SignInViewModel
 import com.captures2024.soongan.core.viewmodel.effect.AppRootSideEffect
+import com.captures2024.soongan.core.viewmodel.effect.SignInSideEffect
 import com.captures2024.soongan.core.viewmodel.intent.AppRootIntent
-import com.captures2024.soongan.feature.signIn.SignInViewModel
-import com.captures2024.soongan.feature.signIn.state.SignInIntent
-import com.captures2024.soongan.feature.signIn.state.SignInSideEffect
+import com.captures2024.soongan.core.viewmodel.intent.SignInIntent
 import com.captures2024.soongan.route.AppRoute
 import com.google.android.gms.auth.api.identity.BeginSignInRequest
 import com.google.android.gms.auth.api.identity.Identity
