+++ app/src/main/kotlin/com/captures2024/soongan/route/AppRoute.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.captures2024.soongan.core.designsystem.component.SoonGanBackground
 import com.captures2024.soongan.core.viewmodel.AppRootViewModel
+import com.captures2024.soongan.core.viewmodel.SignInViewModel
 import com.captures2024.soongan.feature.intro.route.IntroRoute
 import com.captures2024.soongan.feature.main.route.MainRoute
 import com.captures2024.soongan.feature.sign.route.SignRoute
-import com.captures2024.soongan.feature.signIn.SignInViewModel
 import com.captures2024.soongan.ui.AppRootScreen
 
 @Composable
