+++ core/viewmodel/src/main/kotlin/com/captures2024/soongan/core/viewmodel/intent/SignInIntent.kt
@@ -1,4 +1,4 @@
-package com.captures2024.soongan.feature.signIn.state
+package com.captures2024.soongan.core.viewmodel.intent
 
 import com.captures2024.soongan.core.common.base.UIIntent
