+++ feature/sign/src/main/kotlin/com/captures2024/soongan/feature/sign/navigation/SignRouteNavhost.kt
@@ -12,8 +12,8 @@ import com.captures2024.soongan.core.navigator.screen.sign.NicknameNavigator
 import com.captures2024.soongan.core.navigator.screen.sign.PrivacyPolicyNavigator
 import com.captures2024.soongan.core.navigator.screen.sign.SignInNavigator
 import com.captures2024.soongan.core.navigator.screen.sign.TermsOfUseNavigator
+import com.captures2024.soongan.core.viewmodel.SignInViewModel
 import com.captures2024.soongan.feature.privacypolicy.navigation.privacyPolicy
-import com.captures2024.soongan.feature.signIn.SignInViewModel
 import com.captures2024.soongan.feature.signIn.navigation.signIn
 import com.captures2024.soongan.feature.signUp.navigation.signUp
 import com.captures2024.soongan.feature.termsofuse.navigation.termsOfUse
