+++ feature/signIn/src/main/kotlin/com/captures2024/soongan/feature/signIn/navigation/SignInNavigation.kt
@@ -1,10 +1,9 @@
 package com.captures2024.soongan.feature.signIn.navigation
 
 import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.captures2024.soongan.core.navigator.screen.sign.SignInNavigator
-import com.captures2024.soongan.feature.signIn.SignInViewModel
+import com.captures2024.soongan.core.viewmodel.SignInViewModel
 import com.captures2024.soongan.feature.signIn.route.SignInRoute
